backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Marshmallow==1.2.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5
email-validator==2.1.0
cryptography